requirements.txt
Django
PyJWT==2.9.0
numpy==2.2.2